+++ docs/theme/js/tempusdominus-bootstrap-3.js
@@ -1483,7 +1483,7 @@ var DateTimePicker = function ($) {
 }(jQuery);
 //noinspection JSUnusedGlobalSymbols
 /* global DateTimePicker */
-var TempusDominusBootstrap3 = function ($) {
+var TempusDominusBootstrap4 = function ($) {
     // eslint-disable-line no-unused-vars
     var JQUERY_NO_CONFLICT = $.fn[DateTimePicker.NAME],
         verticalModes = ['top', 'bottom', 'auto'],
@@ -1509,11 +1509,11 @@ var TempusDominusBootstrap3 = function ($) {
         return $selector;
     };
 
-    var TempusDominusBootstrap3 = function (_DateTimePicker) {
-        _inherits(TempusDominusBootstrap3, _DateTimePicker);
+    var TempusDominusBootstrap4 = function (_DateTimePicker) {
+        _inherits(TempusDominusBootstrap4, _DateTimePicker);
 
-        function TempusDominusBootstrap3(element, options) {
-            _classCallCheck2(this, TempusDominusBootstrap3);
+        function TempusDominusBootstrap4(element, options) {
+            _classCallCheck2(this, TempusDominusBootstrap4);
 
             var _this = _possibleConstructorReturn(this, _DateTimePicker.call(this, element, options));
 
@@ -1521,7 +1521,7 @@ var TempusDominusBootstrap3 = function ($) {
             return _this;
         }
 
-        TempusDominusBootstrap3.prototype._init = function _init() {
+        TempusDominusBootstrap4.prototype._init = function _init() {
             if (this._element.hasClass('input-group')) {
                 // in case there is more then one 'input-group-addon' Issue #48
                 var datepickerButton = this._element.find('.datepickerbutton');
@@ -1533,14 +1533,14 @@ var TempusDominusBootstrap3 = function ($) {
             }
         };
 
-        TempusDominusBootstrap3.prototype._getDatePickerTemplate = function _getDatePickerTemplate() {
+        TempusDominusBootstrap4.prototype._getDatePickerTemplate = function _getDatePickerTemplate() {
             var headTemplate = $('<thead>').append($('<tr>').append($('<th>').addClass('prev').attr('data-action', 'previous').append($('<span>').addClass(this._options.icons.previous))).append($('<th>').addClass('picker-switch').attr('data-action', 'pickerSwitch').attr('colspan', "" + (this._options.calendarWeeks ? '6' : '5'))).append($('<th>').addClass('next').attr('data-action', 'next').append($('<span>').addClass(this._options.icons.next)))),
                 contTemplate = $('<tbody>').append($('<tr>').append($('<td>').attr('colspan', "" + (this._options.calendarWeeks ? '8' : '7'))));
 
             return [$('<div>').addClass('datepicker-days').append($('<table>').addClass('table-condensed').append(headTemplate).append($('<tbody>'))), $('<div>').addClass('datepicker-months').append($('<table>').addClass('table-condensed').append(headTemplate.clone()).append(contTemplate.clone())), $('<div>').addClass('datepicker-years').append($('<table>').addClass('table-condensed').append(headTemplate.clone()).append(contTemplate.clone())), $('<div>').addClass('datepicker-decades').append($('<table>').addClass('table-condensed').append(headTemplate.clone()).append(contTemplate.clone()))];
         };
 
-        TempusDominusBootstrap3.prototype._getTimePickerMainTemplate = function _getTimePickerMainTemplate() {
+        TempusDominusBootstrap4.prototype._getTimePickerMainTemplate = function _getTimePickerMainTemplate() {
             var topRow = $('<tr>'),
                 middleRow = $('<tr>'),
                 bottomRow = $('<tr>');
@@ -1617,7 +1617,7 @@ var TempusDominusBootstrap3 = function ($) {
             return $('<div>').addClass('timepicker-picker').append($('<table>').addClass('table-condensed').append([topRow, middleRow, bottomRow]));
         };
 
-        TempusDominusBootstrap3.prototype._getTimePickerTemplate = function _getTimePickerTemplate() {
+        TempusDominusBootstrap4.prototype._getTimePickerTemplate = function _getTimePickerTemplate() {
             var hoursView = $('<div>').addClass('timepicker-hours').append($('<table>').addClass('table-condensed')),
                 minutesView = $('<div>').addClass('timepicker-minutes').append($('<table>').addClass('table-condensed')),
                 secondsView = $('<div>').addClass('timepicker-seconds').append($('<table>').addClass('table-condensed')),
@@ -1636,7 +1636,7 @@ var TempusDominusBootstrap3 = function ($) {
             return ret;
         };
 
-        TempusDominusBootstrap3.prototype._getToolbar = function _getToolbar() {
+        TempusDominusBootstrap4.prototype._getToolbar = function _getToolbar() {
             var row = [];
             if (this._options.showTodayButton) {
                 row.push($('<td>').append($('<a>').attr({
@@ -1665,7 +1665,7 @@ var TempusDominusBootstrap3 = function ($) {
             return $('<table>').addClass('table-condensed').append($('<tbody>').append($('<tr>').append(row)));
         };
 
-        TempusDominusBootstrap3.prototype._getTemplate = function _getTemplate() {
+        TempusDominusBootstrap4.prototype._getTemplate = function _getTemplate() {
             var template = $('<div>').addClass('bootstrap-datetimepicker-widget dropdown-menu'),
                 dateView = $('<div>').addClass('datepicker').append(this._getDatePickerTemplate()),
                 timeView = $('<div>').addClass('timepicker').append(this._getTimePickerTemplate()),
@@ -1713,7 +1713,7 @@ var TempusDominusBootstrap3 = function ($) {
             return template.append(content);
         };
 
-        TempusDominusBootstrap3.prototype._place = function _place() {
+        TempusDominusBootstrap4.prototype._place = function _place() {
             var position = (this.component || this._element).position(),
                 offset = (this.component || this._element).offset();
             var vertical = this._options.widgetPositioning.vertical,
@@ -1782,7 +1782,7 @@ var TempusDominusBootstrap3 = function ($) {
             });
         };
 
-        TempusDominusBootstrap3.prototype._fillDow = function _fillDow() {
+        TempusDominusBootstrap4.prototype._fillDow = function _fillDow() {
             var row = $('<tr>'),
                 currentDate = this._viewDate.clone().startOf('w').startOf('d');
 
@@ -1797,7 +1797,7 @@ var TempusDominusBootstrap3 = function ($) {
             this.widget.find('.datepicker-days thead').append(row);
         };
 
-        TempusDominusBootstrap3.prototype._fillMonths = function _fillMonths() {
+        TempusDominusBootstrap4.prototype._fillMonths = function _fillMonths() {
             var spans = [],
                 monthsShort = this._viewDate.clone().startOf('y').startOf('d');
             while (monthsShort.isSame(this._viewDate, 'y')) {
@@ -1807,7 +1807,7 @@ var TempusDominusBootstrap3 = function ($) {
             this.widget.find('.datepicker-months td').empty().append(spans);
         };
 
-        TempusDominusBootstrap3.prototype._updateMonths = function _updateMonths() {
+        TempusDominusBootstrap4.prototype._updateMonths = function _updateMonths() {
             var monthsView = this.widget.find('.datepicker-months'),
                 monthsViewHeader = monthsView.find('th'),
                 months = monthsView.find('tbody').find('span');
@@ -1840,7 +1840,7 @@ var TempusDominusBootstrap3 = function ($) {
             });
         };
 
-        TempusDominusBootstrap3.prototype._updateYears = function _updateYears() {
+        TempusDominusBootstrap4.prototype._updateYears = function _updateYears() {
             var yearsView = this.widget.find('.datepicker-years'),
                 yearsViewHeader = yearsView.find('th'),
                 startYear = this._viewDate.clone().subtract(5, 'y'),
@@ -1871,7 +1871,7 @@ var TempusDominusBootstrap3 = function ($) {
             yearsView.find('td').html(html);
         };
 
-        TempusDominusBootstrap3.prototype._updateDecades = function _updateDecades() {
+        TempusDominusBootstrap4.prototype._updateDecades = function _updateDecades() {
             var decadesView = this.widget.find('.datepicker-decades'),
                 decadesViewHeader = decadesView.find('th'),
                 startDecade = moment({ y: this._viewDate.year() - this._viewDate.year() % 100 - 1 }),
@@ -1910,7 +1910,7 @@ var TempusDominusBootstrap3 = function ($) {
             decadesViewHeader.eq(1).text(startedAt.year() + "1-" + startDecade.year());
         };
 
-        TempusDominusBootstrap3.prototype._fillDate = function _fillDate() {
+        TempusDominusBootstrap4.prototype._fillDate = function _fillDate() {
             var daysView = this.widget.find('.datepicker-days'),
                 daysViewHeader = daysView.find('th'),
                 html = [];
@@ -1980,7 +1980,7 @@ var TempusDominusBootstrap3 = function ($) {
             this._updateDecades();
         };
 
-        TempusDominusBootstrap3.prototype._fillHours = function _fillHours() {
+        TempusDominusBootstrap4.prototype._fillHours = function _fillHours() {
             var table = this.widget.find('.timepicker-hours table'),
                 currentHour = this._viewDate.clone().startOf('d'),
                 html = [];
@@ -2000,7 +2000,7 @@ var TempusDominusBootstrap3 = function ($) {
             table.empty().append(html);
         };
 
-        TempusDominusBootstrap3.prototype._fillMinutes = function _fillMinutes() {
+        TempusDominusBootstrap4.prototype._fillMinutes = function _fillMinutes() {
             var table = this.widget.find('.timepicker-minutes table'),
                 currentMinute = this._viewDate.clone().startOf('h'),
                 html = [],
@@ -2018,7 +2018,7 @@ var TempusDominusBootstrap3 = function ($) {
             table.empty().append(html);
         };
 
-        TempusDominusBootstrap3.prototype._fillSeconds = function _fillSeconds() {
+        TempusDominusBootstrap4.prototype._fillSeconds = function _fillSeconds() {
             var table = this.widget.find('.timepicker-seconds table'),
                 currentSecond = this._viewDate.clone().startOf('m'),
                 html = [];
@@ -2036,7 +2036,7 @@ var TempusDominusBootstrap3 = function ($) {
             table.empty().append(html);
         };
 
-        TempusDominusBootstrap3.prototype._fillTime = function _fillTime() {
+        TempusDominusBootstrap4.prototype._fillTime = function _fillTime() {
             var toggle = void 0,
                 newDate = void 0;
             var timeComponents = this.widget.find('.timepicker span[data-time-component]');
@@ -2062,7 +2062,7 @@ var TempusDominusBootstrap3 = function ($) {
             this._fillSeconds();
         };
 
-        TempusDominusBootstrap3.prototype._doAction = function _doAction(e, action) {
+        TempusDominusBootstrap4.prototype._doAction = function _doAction(e, action) {
             if ($(e.currentTarget).is('.disabled')) {
                 return false;
             }
@@ -2297,7 +2297,7 @@ var TempusDominusBootstrap3 = function ($) {
         //public
 
 
-        TempusDominusBootstrap3.prototype.hide = function hide() {
+        TempusDominusBootstrap4.prototype.hide = function hide() {
             var transitioning = false;
             if (!this.widget) {
                 return;
@@ -2339,7 +2339,7 @@ var TempusDominusBootstrap3 = function ($) {
             this._viewDate = this._date.clone();
         };
 
-        TempusDominusBootstrap3.prototype.show = function show() {
+        TempusDominusBootstrap4.prototype.show = function show() {
             var currentMoment = void 0;
             var useCurrentGranularity = {
                 'year': function year(m) {
@@ -2410,14 +2410,14 @@ var TempusDominusBootstrap3 = function ($) {
             });
         };
 
-        TempusDominusBootstrap3.prototype.destroy = function destroy() {
+        TempusDominusBootstrap4.prototype.destroy = function destroy() {
             this.hide();
             //todo doc off?
             this._element.removeData(DateTimePicker.DATA_KEY);
             this._element.removeData('date');
         };
 
-        TempusDominusBootstrap3.prototype.disable = function disable() {
+        TempusDominusBootstrap4.prototype.disable = function disable() {
             this.hide();
             if (this.component && this.component.hasClass('btn')) {
                 this.component.addClass('disabled');
@@ -2427,7 +2427,7 @@ var TempusDominusBootstrap3 = function ($) {
             }
         };
 
-        TempusDominusBootstrap3.prototype.enable = function enable() {
+        TempusDominusBootstrap4.prototype.enable = function enable() {
             if (this.component && this.component.hasClass('btn')) {
                 this.component.removeClass('disabled');
             }
@@ -2436,7 +2436,7 @@ var TempusDominusBootstrap3 = function ($) {
             }
         };
 
-        TempusDominusBootstrap3.prototype.toolbarPlacement = function toolbarPlacement(_toolbarPlacement) {
+        TempusDominusBootstrap4.prototype.toolbarPlacement = function toolbarPlacement(_toolbarPlacement) {
             if (arguments.length === 0) {
                 return this._options.toolbarPlacement;
             }
@@ -2455,7 +2455,7 @@ var TempusDominusBootstrap3 = function ($) {
             }
         };
 
-        TempusDominusBootstrap3.prototype.widgetPositioning = function widgetPositioning(_widgetPositioning) {
+        TempusDominusBootstrap4.prototype.widgetPositioning = function widgetPositioning(_widgetPositioning) {
             if (arguments.length === 0) {
                 return $.extend({}, this._options.widgetPositioning);
             }
@@ -2486,7 +2486,7 @@ var TempusDominusBootstrap3 = function ($) {
             this._update();
         };
 
-        TempusDominusBootstrap3.prototype.widgetParent = function widgetParent(_widgetParent) {
+        TempusDominusBootstrap4.prototype.widgetParent = function widgetParent(_widgetParent) {
             if (arguments.length === 0) {
                 return this._options.widgetParent;
             }
@@ -2509,7 +2509,7 @@ var TempusDominusBootstrap3 = function ($) {
         //static
 
 
-        TempusDominusBootstrap3._jQueryInterface = function _jQueryInterface(option, argument) {
+        TempusDominusBootstrap4._jQueryInterface = function _jQueryInterface(option, argument) {
             return this.each(function () {
                 var data = $(this).data(DateTimePicker.DATA_KEY);
                 if ((typeof option === "undefined" ? "undefined" : _typeof(option)) === 'object') {
@@ -2517,7 +2517,7 @@ var TempusDominusBootstrap3 = function ($) {
                 }
 
                 if (!data) {
-                    data = new TempusDominusBootstrap3($(this), option);
+                    data = new TempusDominusBootstrap4($(this), option);
                     $(this).data(DateTimePicker.DATA_KEY, data);
                 }
 
@@ -2530,7 +2530,7 @@ var TempusDominusBootstrap3 = function ($) {
             });
         };
 
-        return TempusDominusBootstrap3;
+        return TempusDominusBootstrap4;
     }(DateTimePicker);
 
     /**
@@ -2545,13 +2545,13 @@ var TempusDominusBootstrap3 = function ($) {
         if ($target.length === 0) {
             return;
         }
-        TempusDominusBootstrap3._jQueryInterface.call($target, 'toggle');
+        TempusDominusBootstrap4._jQueryInterface.call($target, 'toggle');
     }).on(DateTimePicker.Event.CHANGE, "." + DateTimePicker.ClassName.INPUT, function (event) {
         var $target = getSelectorFromElement($(this));
         if ($target.length === 0) {
             return;
         }
-        TempusDominusBootstrap3._jQueryInterface.call($target, '_change', event);
+        TempusDominusBootstrap4._jQueryInterface.call($target, '_change', event);
     }).on(DateTimePicker.Event.BLUR, "." + DateTimePicker.ClassName.INPUT, function (event) {
         var $target = getSelectorFromElement($(this));
         if ($target.length === 0) {
@@ -2561,19 +2561,19 @@ var TempusDominusBootstrap3 = function ($) {
         if (config._options.debug) {
             return;
         }
-        TempusDominusBootstrap3._jQueryInterface.call($target, 'hide', event);
+        TempusDominusBootstrap4._jQueryInterface.call($target, 'hide', event);
     }).on(DateTimePicker.Event.KEYDOWN, "." + DateTimePicker.ClassName.INPUT, function (event) {
         var $target = getSelectorFromElement($(this));
         if ($target.length === 0) {
             return;
         }
-        TempusDominusBootstrap3._jQueryInterface.call($target, '_keydown', event);
+        TempusDominusBootstrap4._jQueryInterface.call($target, '_keydown', event);
     }).on(DateTimePicker.Event.KEYUP, "." + DateTimePicker.ClassName.INPUT, function (event) {
         var $target = getSelectorFromElement($(this));
         if ($target.length === 0) {
             return;
         }
-        TempusDominusBootstrap3._jQueryInterface.call($target, '_keyup', event);
+        TempusDominusBootstrap4._jQueryInterface.call($target, '_keyup', event);
     }).on(DateTimePicker.Event.FOCUS, "." + DateTimePicker.ClassName.INPUT, function (event) {
         var $target = getSelectorFromElement($(this));
         if ($target.length === 0) {
@@ -2583,15 +2583,15 @@ var TempusDominusBootstrap3 = function ($) {
         if (!config._options.allowInputToggle) {
             return;
         }
-        TempusDominusBootstrap3._jQueryInterface.call($target, config, event);
+        TempusDominusBootstrap4._jQueryInterface.call($target, config, event);
     });
 
-    $.fn[DateTimePicker.NAME] = TempusDominusBootstrap3._jQueryInterface;
-    $.fn[DateTimePicker.NAME].Constructor = TempusDominusBootstrap3;
+    $.fn[DateTimePicker.NAME] = TempusDominusBootstrap4._jQueryInterface;
+    $.fn[DateTimePicker.NAME].Constructor = TempusDominusBootstrap4;
     $.fn[DateTimePicker.NAME].noConflict = function () {
         $.fn[DateTimePicker.NAME] = JQUERY_NO_CONFLICT;
-        return TempusDominusBootstrap3._jQueryInterface;
+        return TempusDominusBootstrap4._jQueryInterface;
     };
 
-    return TempusDominusBootstrap3;
+    return TempusDominusBootstrap4;
 }(jQuery);